% Generated by build script
\name{mfld_groups}
\alias{mfld_groups}
\title{List Manifold groups (topics/categories)}
\usage{
mfld_groups()
}
\arguments{
\item{available_to_anyone}{If TRUE, only public groups (default TRUE)}
}
\value{tibble: id, name, slug, total_members}
\description{List Manifold groups (topics/categories)}
