% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfld.R
\name{mfld_market}
\alias{mfld_market}
\title{Get a single Manifold market by ID}
\usage{
mfld_market(id)
}
\arguments{
\item{id}{Market ID or slug}
}
\value{
tibble: one row with id, question, url, probability, volume,
creator_username, mechanism, description, is_resolved, total_liquidity
}
\description{
Get a single Manifold market by ID
}
