% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfld.R
\name{mfld_markets}
\alias{mfld_markets}
\title{List Manifold Markets prediction markets}
\usage{
mfld_markets(limit = 10, term = NULL)
}
\arguments{
\item{limit}{Max results (default 10, max 1000)}

\item{term}{Search term to filter markets}
}
\value{
tibble: id, question, url, probability, volume, creator_username,
mechanism, created_time, close_time, is_resolved
}
\description{
List Manifold Markets prediction markets
}
