% Generated by build script
\name{mfld_user}
\alias{mfld_user}
\title{Get a Manifold user profile by username}
\usage{
mfld_user()
}
\arguments{
\item{username}{Manifold username}
}
\value{tibble: id, name, username, created_time, balance, total_deposits, profit_cached}
\description{Get a Manifold user profile by username}
